requirements.txt
absl-py==2.4.0
ackermann-msgs==2.0.2
action-msgs==1.2.2
action-tutorials-interfaces==0.20.7
action-tutorials-py==0.20.7
actionlib-msgs==4.9.0
ament-cmake-test==1.3.12
ament-copyright==0.12.14
ament-cppcheck==0.12.14
ament-cpplint==0.12.14
ament-flake8==0.12.14
ament-index-python==1.4.0
ament-lint==0.12.14
ament-lint-cmake==0.12.14
ament-package==0.14.1
ament-pep257==0.12.14
ament-uncrustify==0.12.14
ament-xmllint==0.12.14
angles==1.15.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
builtin-interfaces==1.2.2
catkin-pkg==1.1.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
composition-interfaces==1.2.2
contourpy==1.3.2
control-msgs==4.8.0
controller-manager==2.53.0
controller-manager-msgs==2.53.0
customtkinter==5.2.2
cv-bridge==3.2.1
cycler==0.12.1
darkdetect==0.8.0
demo-nodes-py==0.20.7
diagnostic-msgs==4.9.0
docutils==0.22.4
domain-coordinator==0.10.0
empy==3.3.4
example-interfaces==0.9.3
examples-rclpy-executors==0.15.5
examples-rclpy-minimal-action-client==0.15.5
examples-rclpy-minimal-action-server==0.15.5
examples-rclpy-minimal-client==0.15.5
examples-rclpy-minimal-publisher==0.15.5
examples-rclpy-minimal-service==0.15.5
examples-rclpy-minimal-subscriber==0.15.5
exceptiongroup==1.3.1
fastapi==0.135.1
flatbuffers==25.12.19
fonttools==4.61.1
gazebo-msgs==3.9.0
generate-parameter-library-py==0.6.0
geometry-msgs==4.9.0
h11==0.16.0
idna==3.11
image-geometry==3.2.1
ImageIO==2.37.2
interactive-markers==2.3.2
Jinja2==3.1.6
joblib==1.5.3
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
kiwisolver==1.4.9
lark==1.3.1
laser-geometry==2.4.1
launch==1.0.13
launch-ros==0.19.13
launch-testing==1.0.13
launch-testing-ros==0.19.13
launch-xml==1.0.13
launch-yaml==1.0.13
lazy_loader==0.4
lifecycle-msgs==1.2.2
logging-demo==0.20.7
lynx-interfaces==0.0.0
-e git+https://github.com/Eden-sudo/lynx_calculus_project.git@1b24417a0bdd9b488c3cd7f2c5215804e58ab5b0#egg=lynx_motion_core&subdirectory=build/lynx_motion_core
map-msgs==2.1.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mediapipe==0.10.9
message-filters==4.3.11
nav-msgs==4.9.0
networkx==3.4.2
numpy==2.2.6
opencv-contrib-python==4.13.0.92
osrf-pycommon==2.1.6
packaging==26.0
pcl-msgs==1.0.0
pendulum-msgs==0.20.7
pillow==12.1.0
potracer==0.0.4
protobuf==3.20.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
pyserial==3.5
python-dateutil==2.9.0.post0
python-qt-binding==1.1.3
PyYAML==6.0.3
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.7
rcl-interfaces==1.2.2
rclpy==3.3.19
rcutils==5.1.8
requests==2.32.5
resource-retriever==3.1.3
rmw-dds-common==1.6.0
ros2action==0.18.15
ros2bag==0.15.15
ros2cli==0.18.15
ros2component==0.18.15
ros2controlcli==2.53.0
ros2doctor==0.18.15
ros2interface==0.18.15
ros2launch==0.19.13
ros2lifecycle==0.18.15
ros2multicast==0.18.15
ros2node==0.18.15
ros2param==0.18.15
ros2pkg==0.18.15
ros2plugin==5.1.3
ros2run==0.18.15
ros2service==0.18.15
ros2topic==0.18.15
rosbag2-interfaces==0.15.15
rosbag2-py==0.15.15
rosgraph-msgs==1.2.2
rosidl-adapter==3.1.8
rosidl-cli==3.1.8
rosidl-cmake==3.1.8
rosidl-generator-c==3.1.8
rosidl-generator-cpp==3.1.8
rosidl-generator-py==0.14.6
rosidl-parser==3.1.8
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.4
rosidl-typesupport-fastrtps-cpp==2.2.4
rosidl-typesupport-introspection-c==3.1.8
rosidl-typesupport-introspection-cpp==3.1.8
rpyutils==0.2.2
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.9
rqt-gui-py==1.1.9
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.9
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
scripts==3.9.0
sensor-msgs==4.9.0
sensor-msgs-py==4.9.0
shape-msgs==4.9.0
six==1.17.0
sounddevice==0.5.5
sros2==0.10.7
starlette==0.52.1
statistics-msgs==1.2.2
std-msgs==4.9.0
std-srvs==4.9.0
stereo-msgs==4.9.0
teleop-twist-keyboard==2.4.1
tf2-geometry-msgs==0.25.18
tf2-kdl==0.25.18
tf2-msgs==0.25.18
tf2-py==0.25.18
tf2-ros-py==0.25.18
tf2-sensor-msgs==0.25.18
tf2-tools==0.25.18
threadpoolctl==3.6.0
tifffile==2025.5.10
topic-monitor==0.20.7
trajectory-msgs==4.9.0
turtlesim==1.4.3
typeguard==4.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
unique-identifier-msgs==2.2.1
urllib3==2.6.3
uvicorn==0.42.0
visualization-msgs==4.9.0
xacro==2.1.1
